requirements.txt
absl-py==1.2.0
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
box2d-py==2.3.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
contourpy==1.0.5
cryptography==36.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
executing==1.1.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
flatbuffers==22.9.24
fonttools==4.37.3
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.49.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
install==1.3.5
ipykernel==6.17.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==4.11.2
jwt==1.3.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.3.24
stack-data==0.5.1
swig==4.0.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
